move test to new test case
